Tests/LibWeb/Text/expected/svg-relayout-scroll-frames-preserved.txt

scrollTop=150
hit=probe
PASS
